libchromeos: Prevent starvation in BaseMessageLoop.
When more than once source of events (file descriptor available or
delayed task ready) is available, the MessageLoop has some freedom to
schedule between I/O tasks and delayed tasks. The
base::MessageLoopForIO will not dispatch posted delayed tasks if there
are file descriptors indefinitely available and will prefer to allways
run all the I/O callbacks before running the delayed tasks that are
due, causing what is known as starvation.

In the Chrome context, this is not a big problem since the thread
running the base::MessageLoopForIO normally posts delayed tasks to
other threads. Even with a single CPU, it is sufficient to have a fair
thread scheduler to avoid starvation. In the Chrome OS context of
single threaded daemons doing I/O, this is more important.

This patch tightens the contract on chromeos::MessageLoop API to
prevent starvation. The chromeos::GlibMessageLoop implementation
already ensures this property, so we fix the chromeos::BaseMessageLoop
by posting a task to the message loop when a file descriptor is ready
and stop watching that file descriptor while the task is waiting in the
message loop.

New unittest are included here to verify the same behaviour over both
implementations.

BUG=chromium:419827
TEST=Added unittest.

Change-Id: Ibe0be7ee838dab96b3c94f5bfe4c0fe557eaa101
Reviewed-on: https://chromium-review.googlesource.com/293334
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>