permission_broker: Stop generating firewalld DBus headers.
firewalld is already a build dependency of
chromeos-base/permission_broker, so we should rely on that dependency to
generate and install the headers into the sysroot.

Furthermore, removing this direct pulling will soon become necessary as
we move firewalld out of platform2.

CQ-DEPEND=CL:293634
BUG=chromium:520812
TEST=Trybot

Change-Id: Iaf012b38b1285e6ea97728140b827f972c7d32d9
Reviewed-on: https://chromium-review.googlesource.com/293783
Reviewed-by: Dan Erat <derat@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>