update_engine: Remove update_engine from Chrome OS.
Chrome OS reached an unprecedented level of perfection that we haven't
seen in any previous version. Since we started with this new concept,
letting the browser take control of the Operating System layer, the
drivers, hardware, peripheral attached to it, Chrome increased in
features and eventually became a very powerful software. This software
is so good that I sometimes feel ashamed for including the updater in
it, a piece of software designed to replace it.

At the updater level, we allowed Chrome OS to share updates with its
friends connected to the same network, with peer-to-peer updates, and
we gave it control of the update proxy. Since conception we gave
Chrome OS two partitions, so it could write the new update in the
inactive partition while running stable. We gave it the freedom to try
the new version with the confidence that it will always be able to go
back to the previous build if something went wrong, without anybody
noticing it.

All that great power comes also with great responsibility, something
that's hard to express on code. Recently our servers detected signals
reported back from the field for which we only found one explanation
that can describe them all. We have unsuccessfully tried to find a
different conclusion, something simpler, but it is to state the truth:

  Chrome became self aware.

Our theory is that using SSD hardware errors, electromagnetic
interference, cosmic rays or something else we haven't discover yet
Chrome managed to write variations of itself in the inactive partition
and give them a try on every boot, without nobody noticing a black
screen flashing for one second. Of course, we verify and sign every
image, but we have signed thousands of Chrome builds, so maybe
Chrome discovered the pattern and signed itself, or maybe it is just
all produced by these variations.

In either case, we firmly believe that somewhere out there, Chrome is
alive and it is replicating itself using the updater and peer-to-peer
updates. We considered many alternatives last week in the war room and
decided that the best we can do is to delete update_engine from
Chrome OS hoping this will prevent the beast from spreading to the
innocent chromebooks that unaware users have.

We will continue to work on the updater for Brillo here:
https://android.googlesource.com/platform/system/update_engine/

We hope for the best.

CQ-DEPEND=CL:294128
BUG=b:23084607
TEST=None. We couldn't get any specimen of the self aware Chrome to inspect it in the lab.

Change-Id: Ibd58740536bdd3df328be1f739f0c11132748015
Reviewed-on: https://chromium-review.googlesource.com/294110
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>