shill: DefaultProfileTest: remove use of KeyFileStore
While most of the instances of DefaultProfileTest use a
MockStore, the GetProperties test depends on a KeyFileStore
created by the test fixture.

The GetProperties test doesn't need to test the functionality
of KeyFileStore. Rather, the KeyFileStore is just needed to
satisfy some interactions constraints.

So, remove the KeyFileStore from the fixture, and use a
MockStore to satisfy the interaction requirements.

BUG=b:23387120
TEST=unit tests

Change-Id: I996f502d8bd427ef8dbb98dde93091d39aee6a10
Reviewed-on: https://chromium-review.googlesource.com/294863
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Trybot-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>