trunks: Do not invoke USB clear functions, unless necessary
Invoking these functions in case USB was not initialized seems to be
working fine in the chroor, but causes lock ups when using libftdi in
Ubuntu (as opposed to libftdi1).

Add a filed to the MPSSE structure to keep track of the interface
creation progress. In fact this field was present in the original
usbmpsse package, but was dropped at some point along the way.

TEST=program does not crash on start any more, connects to the UDB
BUG=chrome-os-partner:43025
TEST=manual

Change-Id: Id0f9f327fd8847908e25611cf96818ab6d398273
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295550
Reviewed-by: Bill Richardson <wfrichar@chromium.org>