shill: cellular: fix couple bugs introduced from previous commits
Fix a bug in ModemManager1 where proxy is being reset when disconnected.
Proxy is used for monitoring the service as well, so shouldn't reset it
when service vanished.

Use KeyValueStore's API to retrieve RPCIdentifiers, so that it is not
dependent on the underlying RPC types.

BUG=chromium:517680
TEST=Run network_3GSmokeTest

Change-Id: If5016b1a329dedf232a11f0dbdc592924aa83c83
Reviewed-on: https://chromium-review.googlesource.com/295459
Commit-Ready: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>