shill: KeyFileStore: remove GLib intermediate object
As noted in the previous change, we never make use
of the ability to pass a MockGLib to KeyFileStore.
Further, there should be no future need to do so,
since MockStore should meet any such needs.

Accordingly, update KeyFileStore to call glib functions
directly, rather than via the GLib object.

Note that this CL only changes code inside of KeyFileStore.
We'll update the API, and callers, in a later CL. Likewise
for removing the now-unused code in the GLib class.

BUG=b:23386647
TEST=compile

Change-Id: Ice35d0a2da5dc11f1e65583b092b5079bc6a9156
Reviewed-on: https://chromium-review.googlesource.com/295577
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>