init: Turn Bluetooth on for s3halt
If we're using s3halt we want Bluetooth to be on.  Unfortunately by
the time we get to s3halt then "bluetoothd" has already stopped and
turned off power to Bluetooth.  Let's turn it back on.

See the description in s3halt.conf for details about the tradeoffs
with the current approach.

BUG=chrome-os-partner:44694
TEST=Can wake up from BT iff BT was powered before suspend.
TEST=Hack out BT module and make sure halt still works.

Change-Id: I3ea8e14ba097a740d9c9ba52eac3dbcee875fe5f
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/296427
Reviewed-by: Dan Erat <derat@chromium.org>