libchromeos: Improve default logging when sending asynchronous HTTP requests
When analyzing production logs (with log level = 0/INFO) it is difficult
to judge when HTTP request succeed or fail. With --v=2 enabled, all of
this is provided, but this is not good for production devices and their
logs.

Added request/response tracking messages to LOG(INFO) level to make
log analysis easier.

Bug: b:23791756
Change-Id: Ie3bf48d45d6facb02a6d26176d10a5612d4e683b