shill: change DBus binding file extensions from .xml to .dbus-xml
Add .dbus-xml DBus binding files to shill/dbus_bindings/, which are
duplicates of the .xml files in that folder. Modify shill gyp files
to use these .dbus-xml files when generating DBus adaptors in the
Chrome OS build process.

This change allows both the Android and Chrome OS build systems to
generate DBus adaptors from the same .dbus-xml files. The .xml files
will be deleted in a separate CL so as not to break Chrome OS builds.

BUG: 22956441
TEST: shill compiles successfully using emerge.
TEST: DBus adaptors generated during Android make process.
Change-Id: I5f896ed3569575122bea689116b878750e1c2791