shill: do not persist disable for prohibited technologies
When technologies are disabled or enabled, they should explicitly be
stored in the profile to make the selection persistent across restarts.
Prohibited technologies should be disabled only when the technology is
prohibited.  This should not persist across shill restarts, especially
when the technology is no longer prohibited.  Updated existing unit
test and added new unit test to check for correct persist/nonpersist calls.

Bug: 24142276
TEST=ran unit tests and checked manually with:
 1 - check /var/cache/shill/default.profile for Powered=true on wifi device
 2 - dbus-send --system --print-reply --fixed --dest=org.chromium.flimflam / \
       org.chromium.flimflam.Manager.SetProperty string:ProhibitedTechnologies \
       variant:string:wifi
 3 - dbus-send --system --print-reply --fixed --dest=org.chromium.flimflam / \
       org.chromium.flimflam.Manager.SetProperty string:ProhibitedTechnologies \
       variant:string:""
 4 - check /var/cache/shill/default.profile for Powered=true on wifi device

Change-Id: I286e140641e9be4785a55f300c746e9d93a4b889