shill: Validate file descriptor before passing it to permission_broker over D-Bus
Chrome OS D-Bus generates a fatal error on messages that contains file
descriptor that have not been validated. And that's case when shill
trying to pass a file descriptor to permission_broker for setting
up a firewall rule for VPN.

Fix it by validating the file descriptor before passing it over
to permission_broker through D-Bus.

Bug: None
TEST=Verify device can connect to VPN network using "cros vpn test"
TEST=chrome extension.

Change-Id: I08c325592269952fdf1336fdf1e78af1dd319944