shill: make all calls to NetlinkManager::OnPendingDumpComplete synchronous
Make all calls to NetlinkManager::OnPendingDumpComplete synchronous.
This eliminate edge cases where a synchronous call to
OnPendingDumpComplete() gets executed before message-loop-dispatched
call to the same function, even though the latter was meant to
run first.

Previously, OnPendingDumpComplete() was posted to the message loop
a precaution to prevent shill from blocking on the sending of the
next queued dump message (via RegisterHandlersAndSendMessage).
However, blocking on the sending of the next message should is not
an issue as the handler-registration and message-sending routines
are minimal and straightforward. Moreover, we already block
shill on the sending of netlink messages in SendOrPostMessage().

BUG=chrome-os-partner:45626
TEST=Compile shill and run unit tests.

Change-Id: Ifdd302527f95be2d63db9af3292429168526489f