shill: dbus: release D-Bus object upon proxy destruction
Currently, the generated Chrome OS D-Bus proxies are leaking
memory/resources, since it never remove the underlying object
proxy from the bus upon destruction.  This causes a problem
when we frequently create/destroy proxies (ChromeosSupplicantBSSProxy),
where the bus would permenantly fail to add match rules for
new object proxies once the limit for match rules is reached.
Without match rules, object proxies will not be able to receive
any signals from the bus.

Previously, we tried to resolve this issue by releasing D-Bus
object upon proxy destruction for all proxies (CL:171310). This causes
problem for proxies with well known path. When multiple instances
of such proxies are created, detructing the first one will remove
the underlying D-Bus object on the bus. SupplicantProcessProxy is
one of them, it is created by both Ethernet (for wired 802.1x) and
WiFi.

So to work around this issue, only release the D-Bus objects for
proxies with unique paths.  The memory/resource leaks from the
proxies with well known paths are very insignificant compare to the
proxies that were created dynamically/on-demand.

Bug: 23560086
TEST=Run wifi_matfunc test
TEST=Run platform_ExternalUsbPeripherals.set1 on a DUT where shill would
     crash with the previous fix CL:171310
TEST=Run network_WiMaxSmoke test

Change-Id: I2808e6b80f3753432cb36e51f38d6dcbd6d880ed