shill: cancel resend dump message task on dump reply timeout
Cancel the delayed message loop task to resend a pending
dump message when we timeout waiting for replies to the
pending dump message. Achieve this by wrapping this task
in a CancelableClosure, and canceling it in
OnPendingDumpComplete(), which is called by
OnPendingDumpTimeout().

This prevents NetlinkManager::ResendPendingDumpMessage from
running for a dump message that has already timed out.

While there, add an extra check in
NetlinkManager::ResendPendingDumpMessage to make sure we
do not attempt to send a non-null message.

BUG: 24300540
Change-Id: Ib34f6083da249dd8bec5e85d6c35484759bd9928
TEST: Compile shill and run unit tests.