shill: fix synchronous shutdown
When the system has a cellular modem, shill takes
care to wait for the cellular connection to be torn
down, before exiting. To do so, shill waits for some
asynchronous termination actions to complete.

In order to wait for these termination actions,
ChromeosDBusDaemon does the following:
1. Call ChromeosDaemon::Quit(), with a completion callback
   which will break out of the message loop.
2. Let ChromeosDaemon::Quit() initiate any appropriate
   asynchronous calls.
3. Run the message loop.
4. Expect ChromeosDaemon to invoke the callback from 1,
   to break out the message loop.

This scheme works fine when we do have asynchronous
work. But we run into problems when we do _not_ have
asynchronous work. The problem is that, in this case,
ChromeosDaemon::Quit() invokes the completion callback
synchronously.

Because the completion callback executes synchronously,
we attempt to break out of the message loop before the
message loop is actually running. When the message loop
does start running, it keeps running indefinitely.
That's because the completion callback has already run.

To fix this, we have ChromeosDaemon::Quit() indicate
whether or not termination can proceed synchronously.
If termination can proceed synchronously,
ChromeosDBusDaemon::OnShutdown() does not attempt
to run the message loop.

While there, we modify ChromeosDaemon to _not_
invoke the completion callback, if all work completed
synchronously. This part of the change avoids the
following log message:
  Message loop not running, ignoring BreakLoop().

Arguably, this second change isn't necessary. The log
message is a VERBOSE1, which is normally off. Nonetheless,
I think skipping the callback in the synchronous case makes
the logic easier to understand.

Bug: None
BUG=chrome-os-partner:45309
TEST=unit tests, manual

Manual test
-----------
- On a device _without_ a cellular modem, run
     date && restart shill > /dev/null && date
  The differences between the two timestamps should be small
  (approx. 1 second).
- On a device _with_ a cellular modem, run
     (ff_debug --level -1 && ff_debug +daemon) > /dev/null \
     && date && restart shill > /dev/null && date \
     && egrep 'Will wait' /var/log/net.log
  You should see 'Will wait for termination actions to complete'

Change-Id: Ieeb7a66a29a49f7c0b133044462a89bb1dee1ffc